+++ packages/codemirror-block-editor/src/extension/handle-change-within-block-level.ts
@@ -1,16 +1,15 @@
 import { EditorState } from "@codemirror/basic-setup";
 import { ChangeSpec } from "@codemirror/state";
-import { findBlockLevelOfLine } from "./find-block-level-of-line";
+import { findBlockLevelOfLineNumberInState } from "./find-block-level-of-line";
 
 export const handleChangeWithinBlockLevel = EditorState.transactionFilter.of(
   (transaction) => {
     const { doc } = transaction.startState;
     const changes: ChangeSpec[] = [];
     transaction.changes.iterChanges((fromA, toA, fromB, toB, text) => {
-      console.log("change: ", fromA, toA, fromB, toB, text);
       const fromLine = doc.lineAt(fromA);
       const toLine = doc.lineAt(toA);
-      const fromLevel = findBlockLevelOfLine(
+      const fromLevel = findBlockLevelOfLineNumberInState(
         transaction.startState,
         fromLine.number
       );
@@ -21,6 +20,7 @@ export const handleChangeWithinBlockLevel = EditorState.transactionFilter.of(
         !text.line(1).text.trim().length
       ) {
         // whitespace got inserted at beginning of line
+        // => level increase
         return;
       }
       if (
@@ -29,11 +29,18 @@ export const handleChangeWithinBlockLevel = EditorState.transactionFilter.of(
         toA - fromLine.from < fromLevel
       ) {
         // whitespace deleted in indentation
+        // => level decrease
         return;
       }
 
       if (fromA - fromLine.from < fromLevel) {
-        const deleteLineBreakOfPreviousLine = fromLine.number === 1 ? 0 : 1;
+        const deleteLineBreakOfPreviousLine =
+          // first line
+          fromLine.number === 1 ||
+          //
+          toA === fromLine.to + 1
+            ? 0
+            : 1;
         changes.push({
           from: fromLine.from - deleteLineBreakOfPreviousLine,
           to: fromLine.from + fromLevel,