+++ packages/codemirror-block-editor/src/index.test.ts
@@ -1,7 +1,5 @@
-import { expect } from "chai";
-
 describe("Index", () => {
   it("should pass", () => {
-    expect(true).to.be.true;
+    expect(true).toEqual(true);
   });
 });